Update README.md
The app usually starts with a user-friendly interface created using HTML and styled with CSS. This interface displays weather-related information such as temperature, weather conditions, location, and more. The core of the app is an index.js server that handles user requests, interacts with external weather APIs, and sends data to the front end. To provide accurate weather data, the app often connects to external weather API OpenWeatherMap.